requirements.txt
flask==3.0.2
mysql-connector-python==8.3.0
chromadb==0.4.22
python-dotenv==1.0.1
werkzeug>=3.0.0
PyPDF2==3.0.1
python-docx==1.0.1
transformers==4.36.2
torch>=2.6.0
sentence-transformers==2.3.1
requests==2.31.0
bcrypt==4.1.2
python-jose==3.3.0
pandas>=2.2.0
numpy>=1.26.4
openpyxl==3.1.2
openai==1.12.0